+++ src/elements/db/ElementQueryInterface.php
@@ -286,11 +286,11 @@ interface ElementQueryInterface extends QueryInterface, ArrayAccess, Arrayable, 
      *     ->one();
      * ```
      *
-     * @param bool $value The property value (defaults to true)
+     * @param bool|null $value The property value (defaults to true)
      * @return self self reference
      * @since 3.2.0
      */
-    public function revisions(bool $value = true): self;
+    public function revisions(?bool $value = true): self;
 
     /**
      * Narrows the query results based on the {elements}’ revision’s ID (from the `revisions` table).